requirements.txt
# Chess Engine Training Requirements
# Optimized for H100/A100 GPU training

# Core ML
torch>=2.0.0
numpy>=1.24.0

# Chess
python-chess>=1.999

# Data processing
zstandard>=0.21.0
requests>=2.28.0
tqdm>=4.65.0

# Logging & visualization
tensorboard>=2.12.0

# Benchmarking
stockfish>=3.28.0

# Optional: Multi-processing
multiprocessing-logging>=0.3.4
